% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{complete_data}
\alias{complete_data}
\title{Data Imputation}
\usage{
complete_data(dat, method = "kNN")
}
\arguments{
\item{dat}{metabolomic data. Should contain \code{Compound} column.}

\item{method}{imputation method. See details for more information.}
}
\value{
completed data frame.
}
\description{
This function completes the metabolomic data using four
approaches of imputation.
}
