% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{easyR_gui}
\alias{easyR_gui}
\title{easyR Graphical User Interface}
\usage{
easyR_gui(port = getOption("shiny.port"))
}
\arguments{
\item{port}{The TCP port. See \code{\link[shiny]{runApp}}.}
}
\description{
Launches graphical user interface.
}
\section{Warning }{
 Any ad-blocking software may cause malfunctions.
}

